delete (engine-ws): removed deprecated incomplete WebService implementation from core engine, fixes CAM-1603
