+++ engine/pom.xml
@@ -17,10 +17,6 @@
       <artifactId>commons-email</artifactId>
     </dependency>
     <dependency>
-      <groupId>commons-lang</groupId>
-      <artifactId>commons-lang</artifactId>
-    </dependency>
-    <dependency>
       <!-- required for building with JDK 5 -->
       <groupId>org.livetribe</groupId>
       <artifactId>livetribe-jsr223</artifactId>
@@ -74,11 +70,6 @@
     <dependency>
       <groupId>com.h2database</groupId>
       <artifactId>h2</artifactId>
-      <scope>test</scope>
-    </dependency>
-    <dependency>
-      <groupId>org.mockito</groupId>
-      <artifactId>mockito-core</artifactId>
       <scope>test</scope>
     </dependency>
     <dependency>
@@ -174,7 +165,6 @@
           javax.xml*;resolution:=optional,
           javax.mail*;resolution:=optional,
           org.apache.catalina*;resolution:=optional,
-          org.apache.commons.lang;resolution:=optional,
           org.apache.commons.mail;resolution:=optional,
           org.apache.tools.ant*;resolution:=optional,
           org.apache.xerces*;resolution:=optional,
