--- engine/src/main/java/org/camunda/bpm/engine/impl/bpmn/data/StructureInstance.java
@@ -1,35 +1,0 @@
-/* Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- * 
- *      http://www.apache.org/licenses/LICENSE-2.0
- * 
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.camunda.bpm.engine.impl.bpmn.data;
-
-/**
- * An instance of {@link StructureDefinition}
- * 
- * @author Esteban Robles Luna
- */
-public interface StructureInstance {
-
-  /**
-   * Converts this structure instance into an array
-   * 
-   * @return this structure as an array
-   */
-  Object[] toArray();
-
-  /**
-   * Loads this structure from array
-   * 
-   * @param array the array where this structure loads data
-   */
-  void loadFrom(Object[] array);
-}
