--- engine/src/main/java/org/camunda/bpm/engine/impl/webservice/SyncWebServiceClientFactory.java
@@ -1,29 +1,0 @@
-/* Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- * 
- *      http://www.apache.org/licenses/LICENSE-2.0
- * 
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.camunda.bpm.engine.impl.webservice;
-
-/**
- * A factory for {@link SyncWebServiceClient}
- * 
- * @author Esteban Robles Luna
- */
-public interface SyncWebServiceClientFactory {
-
-  /**
-   * Creates a {@link SyncWebServiceClient} from a WSDL
-   * 
-   * @param wsdl the WSDL to be used in the {@link SyncWebServiceClient}
-   * @return the new instance of {@link SyncWebServiceClient}
-   */
-  SyncWebServiceClient create(String wsdl);
-}
